Add mitigation for failure when a CU cannot be parsed
Currently, when a compilation unit can not be parsed at
all, the analysis process will be terminated with an
exception and the build will fail.

With this change, we first try to fall back to good old
line-by-line parsing when full-cu-parsing fails. Anyway,
if any failure occurs during parsing of a file, analysis
will not terminate but a warning will be added to the
analysis result and printed at the end.

A warning alone does not fail the build. Build will only fail
if at least one banned import was found