[mvn] Make release use more secure checksums
maven defaults to sha-1.  This will ensure it releases more strict checksum.  See that most are using this, typicalyl you can set this on release plugin directly, this should cover it overwise, if downloading, maven 3.9.x requires additional flag to pull.  I've added that here but that is not necessary on release plugin if setting that one specifically.